refactor: split argparse and run_pipline
